std 타입으로 변경

+++ src/geometry/polygon2.h
@@ -17,8 +17,8 @@ struct polygon2
     std::array<vec2, MAX_POLYGON_VERTICES> normals{};
     vec2 centroid{};
     float radius = 0.0f;
-    std::size_t vertexCount = 0;
-};
+    std::int32_t vertexCount = 0;
+}; // sizeof: 144 bytes
 
 polygon2 MakeBox( const vec2& halfExtents );
